モデルのリファクタリング
