+++ app/src/main/kotlin/com/numero/itube/api/response/ChannelResponse.kt
@@ -1,6 +1,8 @@
 package com.numero.itube.api.response
 
-import com.numero.itube.model.Channel
+import com.numero.itube.model.PageInfo
+import com.numero.itube.model.Thumbnail
+import com.squareup.moshi.Json
 import se.ansman.kotshi.JsonSerializable
 import java.io.Serializable
 
@@ -14,8 +16,27 @@ data class ChannelResponse(
 ) : Serializable {
 
     @JsonSerializable
-    data class PageInfo(
-            val totalResults: Int,
-            val resultsPerPage: Int
-    ) : Serializable
+    data class Channel(
+            val kind: String,
+            val etag: String,
+            val id: String,
+            val snippet: Snippet
+    ) : Serializable {
+
+        @JsonSerializable
+        data class Snippet(
+                val title: String,
+                val description: String,
+                val publishedAt: String,
+                val thumbnails: Thumbnails
+        ) : Serializable
+
+        @JsonSerializable
+        data class Thumbnails(
+                @Json(name = "default")
+                val def: Thumbnail,
+                val medium: Thumbnail,
+                val high: Thumbnail
+        ) : Serializable
+    }
 }
