+++ app/src/main/kotlin/com/numero/itube/api/response/SearchResponse.kt
@@ -1,6 +1,8 @@
 package com.numero.itube.api.response
 
-import com.numero.itube.model.Video
+import com.numero.itube.model.PageInfo
+import com.numero.itube.model.Thumbnail
+import com.squareup.moshi.Json
 import se.ansman.kotshi.JsonSerializable
 import java.io.Serializable
 
@@ -15,8 +17,36 @@ data class SearchResponse(
 ) : Serializable {
 
     @JsonSerializable
-    data class PageInfo(
-            val totalResults: Int,
-            val resultsPerPage: Int
-    ) : Serializable
+    data class Video(
+            val kind: String,
+            val etag: String,
+            val id: Id,
+            val snippet: Snippet
+    ) : Serializable {
+
+        @JsonSerializable
+        data class Id(
+                val kind: String,
+                val videoId: String
+        ) : Serializable
+
+        @JsonSerializable
+        data class Snippet(
+                val publishedAt: String,
+                val channelId: String,
+                val title: String,
+                val description: String,
+                val thumbnails: Thumbnails,
+                val channelTitle: String,
+                val liveBroadcastContent: String
+        ) : Serializable
+
+        @JsonSerializable
+        data class Thumbnails(
+                @Json(name = "default")
+                val def: Thumbnail,
+                val medium: Thumbnail,
+                val high: Thumbnail
+        ) : Serializable
+    }
 }
