+++ app/src/main/kotlin/com/numero/itube/contract/DetailContract.kt
@@ -1,14 +1,14 @@
 package com.numero.itube.contract
 
-import com.numero.itube.model.Channel
-import com.numero.itube.model.VideoDetail
+import com.numero.itube.api.response.ChannelResponse
+import com.numero.itube.api.response.VideoDetailResponse
 import com.numero.itube.presenter.IPresenter
 import com.numero.itube.view.IView
 
 interface DetailContract {
 
     interface View : IView<Presenter> {
-        fun showVideoDetail(videoDetail: VideoDetail, channel: Channel)
+        fun showVideoDetail(videoDetail: VideoDetailResponse.VideoDetail, channel: ChannelResponse.Channel)
 
         fun showErrorMessage(e: Throwable?)
