+++ app/src/main/kotlin/com/numero/itube/contract/RelativeContract.kt
@@ -1,13 +1,13 @@
 package com.numero.itube.contract
 
-import com.numero.itube.model.Video
+import com.numero.itube.api.response.SearchResponse
 import com.numero.itube.presenter.IPresenter
 import com.numero.itube.view.IView
 
 interface RelativeContract {
 
     interface View : IView<Presenter> {
-        fun showVideoList(videoList: List<Video>)
+        fun showVideoList(videoList: List<SearchResponse.Video>)
 
         fun showErrorMessage(e: Throwable?)
