+++ app/src/main/kotlin/com/numero/itube/model/PageInfo.kt
@@ -1,0 +1,10 @@
+package com.numero.itube.model
+
+import se.ansman.kotshi.JsonSerializable
+import java.io.Serializable
+
+@JsonSerializable
+data class PageInfo(
+        val totalResults: Int,
+        val resultsPerPage: Int
+) : Serializable
