+++ app/src/main/kotlin/com/numero/itube/model/Thumbnail.kt
@@ -1,0 +1,11 @@
+package com.numero.itube.model
+
+import se.ansman.kotshi.JsonSerializable
+import java.io.Serializable
+
+@JsonSerializable
+data class Thumbnail(
+        val url: String,
+        val width: Int,
+        val height: Int
+) : Serializable
