+++ app/src/main/kotlin/com/numero/itube/view/adapter/RelativeVideoListAdapter.kt
@@ -6,21 +6,21 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.numero.itube.R
-import com.numero.itube.model.Video
+import com.numero.itube.api.response.SearchResponse
 import kotlinx.android.extensions.LayoutContainer
 import kotlinx.android.synthetic.main.view_holder_relative_video.*
 
 class RelativeVideoListAdapter : RecyclerView.Adapter<RelativeVideoListAdapter.VideoViewHolder>() {
 
-    var videoList: List<Video> = listOf()
+    var videoList: List<SearchResponse.Video> = listOf()
         set(value) {
             field = value
             notifyDataSetChanged()
         }
 
-    private var onItemClickListener: ((video: Video) -> Unit)? = null
+    private var onItemClickListener: ((video: SearchResponse.Video) -> Unit)? = null
 
-    fun setOnItemClickListener(listener: ((video: Video) -> Unit)) {
+    fun setOnItemClickListener(listener: ((video: SearchResponse.Video) -> Unit)) {
         onItemClickListener = listener
     }
 
@@ -40,7 +40,7 @@ class RelativeVideoListAdapter : RecyclerView.Adapter<RelativeVideoListAdapter.V
 
     class VideoViewHolder(override val containerView: View) : RecyclerView.ViewHolder(containerView), LayoutContainer {
 
-        fun setVideo(video: Video) {
+        fun setVideo(video: SearchResponse.Video) {
             titleTextView.text = video.snippet.title
             Glide.with(itemView.context).load(video.snippet.thumbnails.high.url).into(thumbnailImageView)
             //.diskCacheStrategy(DiskCacheStrategy.NONE)
